remove punctionation from words with regex
